update(tests) ordered tests and refactored code
