+++ test/seed.spec.js
@@ -1,0 +1,11 @@
+'use strict'
+/* global describe beforeEach it */
+
+const seed = require('./../script/seed')
+
+describe('▒▒▒ Script tests ▒▒▒', () => {
+
+	describe('seed script', () => {
+	  it('completes successfully', seed)
+	})
+})
